% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday-movies.R
\name{movie_die_hard}
\alias{movie_die_hard}
\title{Clearly the best holiday movie: \emph{Die Hard}}
\usage{
movie_die_hard()
}
\description{
Clearly the best holiday movie: \emph{Die Hard}
}
\examples{
movie_home_alone()

}
