% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday-movies.R
\name{movie_home_alone}
\alias{movie_home_alone}
\title{Clearly the best holiday movie: \emph{Home Alone}}
\usage{
movie_home_alone()
}
\value{
A line from the best holiday movie of all time, \emph{Home Alone}
}
\description{
Clearly the best holiday movie: \emph{Home Alone}
}
\examples{
movie_home_alone()

}
