% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday-movies.R
\name{movie_wonderful_life}
\alias{movie_wonderful_life}
\title{Clearly the best holiday movie: \emph{It's a Wonderful Life}}
\usage{
movie_wonderful_life()
}
\value{
A line from the best holiday movie of all time, \emph{It's a Wonderful Life}
}
\description{
Clearly the best holiday movie: \emph{It's a Wonderful Life}
}
\examples{
movie_wonderful_life()

}
